+++ app/monitor/gforce.py
@@ -16,8 +16,8 @@ class GForceMonitor(Monitor):
 
     def __init__(
         self,
-        x_pos: int,
-        y_pos: int,
+        x_pos: float,
+        y_pos: float,
     ) -> None:
         super().__init__(x_pos, y_pos)
 
@@ -49,9 +49,9 @@ class GForceMonitor(Monitor):
         ) if config.GForceMonitor.slip_ratio_enabled else None
 
     @property
-    def width(self) -> int: return self._width
+    def width(self) -> float: return self._width
     @property
-    def height(self) -> int: return self._height
+    def height(self) -> float: return self._height
 
     def _render_gforce_square_dot(self) -> None:
         x_accG, _, z_accG = ac_api[CS.AccG].wma()
